requirements.txt
Flask
ebooklib
newspaper3k
requests
Pillow
lxml_html_clean
gunicorn
